adk-samples/requirements.txt
fastapi
pydantic
logging
google-cloud-pubsub